[error] moved eth utils errors to Error.ts file
