+++ src/Errors.ts
@@ -1,0 +1,9 @@
+export class EthKeyAlreadyExist extends Error{}
+
+export class EthKeyDoesNotExist extends Error{}
+
+export class InvalidPrivateKey extends Error{}
+
+export class InvalidAddress extends Error{}
+
+export class InvalidPublicKey extends Error{}
